% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_extras.R
\name{add_unique_edges}
\alias{add_unique_edges}
\title{Draw a set of additional ties in one layer}
\usage{
add_unique_edges(
  nodes,
  n_edges,
  layer,
  seed,
  same_dept_bias = 0.7,
  same_legacy_bias = NULL
)
}
\arguments{
\item{nodes}{Actors eligible to be joined.}

\item{n_edges}{Number of pairs to draw.}

\item{layer}{The layer the ties belong to.}

\item{seed}{Integer seed.}

\item{same_dept_bias}{Weight given to same-department pairs, in \verb{[0, 1]}.}

\item{same_legacy_bias}{Optional weight for the legacy split.}
}
\value{
An edge tibble in the standard schema.
}
\description{
Draw a set of additional ties in one layer
}
\keyword{internal}
