% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{BASELINE_MODELS}
\alias{BASELINE_MODELS}
\title{Baseline reference topologies}
\format{
A character vector of baseline model identifiers.
}
\usage{
BASELINE_MODELS
}
\description{
Baseline reference topologies
}
