% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{default_evolution}
\alias{default_evolution}
\title{Default tie turnover settings}
\usage{
default_evolution()
}
\value{
A list of evolution settings.
}
\description{
Default tie turnover settings
}
\keyword{internal}
