% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_output_dir}
\alias{default_output_dir}
\title{Default output directory}
\usage{
default_output_dir()
}
\value{
Path to the default output directory.
}
\description{
A session temporary directory. Exports never touch the user's file space
unless an explicit \code{output_root} is supplied.
}
