% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{degree_to_probability}
\alias{degree_to_probability}
\title{Tie probability giving a target mean degree}
\usage{
degree_to_probability(n, mean_degree)
}
\arguments{
\item{n}{Number of nodes.}

\item{mean_degree}{Target mean degree.}
}
\value{
A probability on \verb{[0, 1]}.
}
\description{
Tie probability giving a target mean degree
}
\keyword{internal}
