% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_template}
\alias{get_template}
\title{Retrieve one industry template}
\usage{
get_template(template_id)
}
\arguments{
\item{template_id}{Template identifier, see \code{\link[=synthona_templates]{synthona_templates()}}.}
}
\value{
A template definition list.
}
\description{
Retrieve one industry template
}
