% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{SYNTHONA_LAYERS}
\alias{SYNTHONA_LAYERS}
\title{Relational layers the protocol can generate}
\format{
A character vector of layer identifiers.
}
\usage{
SYNTHONA_LAYERS
}
\description{
Relational layers the protocol can generate
}
