% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{synthona_registry}
\alias{synthona_registry}
\title{The scenario registry}
\usage{
synthona_registry()
}
\value{
A named list of scenario definitions.
}
\description{
The scenario registry
}
\examples{
names(synthona_registry())
}
