% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{write_corpus}
\alias{write_corpus}
\title{Write a corpus to disk}
\usage{
write_corpus(corpus, output_root = default_output_dir(), ...)
}
\arguments{
\item{corpus}{A list of \code{synthona_dataset} objects.}

\item{output_root}{Directory to write into.}

\item{...}{Passed to \code{\link[=synthona_write]{synthona_write()}}.}
}
\value{
The corpus root path, invisibly.
}
\description{
Write a corpus to disk
}
